% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvt_margLL.R
\name{get.marg.LL}
\alias{get.marg.LL}
\title{True marginal likelihood for MVT example}
\usage{
get.marg.LL(data, hyps)
}
\arguments{
  \item{data}{data} \item{hyps}{hyperpriors}
}
\description{
True marginal likelihood for MVT example
}

